PanelEditor: Update hoverHeader state when changing panel title
