+++ public/app/features/dashboard/components/PanelEditor/getPanelFrameOptions.tsx
@@ -1,23 +1,13 @@
 import React from 'react';
 
-import { SelectableValue } from '@grafana/data';
 import { selectors } from '@grafana/e2e-selectors';
 import { config } from '@grafana/runtime';
-import { VizPanel } from '@grafana/scenes';
 import { DataLinksInlineEditor, Input, RadioButtonGroup, Select, Switch, TextArea } from '@grafana/ui';
-import { VizPanelManager, VizPanelManagerState } from 'app/features/dashboard-scene/panel-edit/VizPanelManager';
-import { VizPanelLinks } from 'app/features/dashboard-scene/scene/PanelLinks';
-import {
-  transformSceneToSaveModel,
-  vizPanelToPanel,
-} from 'app/features/dashboard-scene/serialization/transformSceneToSaveModel';
-import { dashboardSceneGraph } from 'app/features/dashboard-scene/utils/dashboardSceneGraph';
-import { getDashboardSceneFor } from 'app/features/dashboard-scene/utils/utils';
 import { getPanelLinksVariableSuggestions } from 'app/features/panel/panellinks/link_srv';
 
 import { GenAIPanelDescriptionButton } from '../GenAI/GenAIPanelDescriptionButton';
 import { GenAIPanelTitleButton } from '../GenAI/GenAIPanelTitleButton';
-import { RepeatRowSelect, RepeatRowSelect2 } from '../RepeatRowSelect/RepeatRowSelect';
+import { RepeatRowSelect } from '../RepeatRowSelect/RepeatRowSelect';
 
 import { OptionsPaneCategoryDescriptor } from './OptionsPaneCategoryDescriptor';
 import { OptionsPaneItemDescriptor } from './OptionsPaneItemDescriptor';
@@ -190,188 +180,4 @@ export function getPanelFrameCategory(props: OptionPaneRenderProps): OptionsPane
           })
         )
     );
-}
-
-export function getPanelFrameCategory2(
-  panelManager: VizPanelManager,
-  panel: VizPanel,
-  repeat?: string
-): OptionsPaneCategoryDescriptor {
-  const descriptor = new OptionsPaneCategoryDescriptor({
-    title: 'Panel options',
-    id: 'Panel options',
-    isOpenDefault: true,
-  });
-
-  const panelLinksObject = dashboardSceneGraph.getPanelLinks(panel);
-  const links = panelLinksObject?.state.rawLinks ?? [];
-  const dashboard = getDashboardSceneFor(panel);
-
-  return descriptor
-    .addItem(
-      new OptionsPaneItemDescriptor({
-        title: 'Title',
-        value: panel.state.title,
-        popularRank: 1,
-        render: function renderTitle() {
-          return <PanelFrameTitle panel={panel} />;
-        },
-        addon: config.featureToggles.dashgpt && (
-          <GenAIPanelTitleButton
-            onGenerate={(title) => panel.setState({ title })}
-            panel={vizPanelToPanel(panel)}
-            dashboard={transformSceneToSaveModel(dashboard)}
-          />
-        ),
-      })
-    )
-    .addItem(
-      new OptionsPaneItemDescriptor({
-        title: 'Description',
-        value: panel.state.description,
-        render: function renderDescription() {
-          return <DescriptionTextArea panel={panel} />;
-        },
-        addon: config.featureToggles.dashgpt && (
-          <GenAIPanelDescriptionButton
-            onGenerate={(description) => panel.setState({ description })}
-            panel={vizPanelToPanel(panel)}
-          />
-        ),
-      })
-    )
-    .addItem(
-      new OptionsPaneItemDescriptor({
-        title: 'Transparent background',
-        render: function renderTransparent() {
-          return (
-            <Switch
-              value={panel.state.displayMode === 'transparent'}
-              id="transparent-background"
-              onChange={() => {
-                panel.setState({
-                  displayMode: panel.state.displayMode === 'transparent' ? 'default' : 'transparent',
-                });
-              }}
-            />
-          );
-        },
-      })
-    )
-    .addCategory(
-      new OptionsPaneCategoryDescriptor({
-        title: 'Panel links',
-        id: 'Panel links',
-        isOpenDefault: false,
-        itemsCount: links?.length,
-      }).addItem(
-        new OptionsPaneItemDescriptor({
-          title: 'Panel links',
-          render: () => <ScenePanelLinksEditor panelLinks={panelLinksObject ?? undefined} />,
-        })
-      )
-    )
-    .addCategory(
-      new OptionsPaneCategoryDescriptor({
-        title: 'Repeat options',
-        id: 'Repeat options',
-        isOpenDefault: false,
-      })
-        .addItem(
-          new OptionsPaneItemDescriptor({
-            title: 'Repeat by variable',
-            description:
-              'Repeat this panel for each value in the selected variable. This is not visible while in edit mode. You need to go back to dashboard and then update the variable or reload the dashboard.',
-            render: function renderRepeatOptions() {
-              return (
-                <RepeatRowSelect2
-                  id="repeat-by-variable-select"
-                  parent={panel}
-                  repeat={repeat}
-                  onChange={(value?: string) => {
-                    const stateUpdate: Partial<VizPanelManagerState> = { repeat: value };
-                    if (value && !panelManager.state.repeatDirection) {
-                      stateUpdate.repeatDirection = 'h';
-                    }
-                    panelManager.setState(stateUpdate);
-                  }}
-                />
-              );
-            },
-          })
-        )
-        .addItem(
-          new OptionsPaneItemDescriptor({
-            title: 'Repeat direction',
-            showIf: () => !!panelManager.state.repeat,
-            render: function renderRepeatOptions() {
-              const directionOptions: Array<SelectableValue<'h' | 'v'>> = [
-                { label: 'Horizontal', value: 'h' },
-                { label: 'Vertical', value: 'v' },
-              ];
-
-              return (
-                <RadioButtonGroup
-                  options={directionOptions}
-                  value={panelManager.state.repeatDirection ?? 'h'}
-                  onChange={(value) => panelManager.setState({ repeatDirection: value })}
-                />
-              );
-            },
-          })
-        )
-        .addItem(
-          new OptionsPaneItemDescriptor({
-            title: 'Max per row',
-            showIf: () => Boolean(panelManager.state.repeat && panelManager.state.repeatDirection === 'h'),
-            render: function renderOption() {
-              const maxPerRowOptions = [2, 3, 4, 6, 8, 12].map((value) => ({ label: value.toString(), value }));
-              return (
-                <Select
-                  options={maxPerRowOptions}
-                  value={panelManager.state.maxPerRow}
-                  onChange={(value) => panelManager.setState({ maxPerRow: value.value })}
-                />
-              );
-            },
-          })
-        )
-    );
-}
-
-interface ScenePanelLinksEditorProps {
-  panelLinks?: VizPanelLinks;
-}
-
-function ScenePanelLinksEditor({ panelLinks }: ScenePanelLinksEditorProps) {
-  const { rawLinks: links } = panelLinks ? panelLinks.useState() : { rawLinks: [] };
-
-  return (
-    <DataLinksInlineEditor
-      links={links}
-      onChange={(links) => panelLinks?.setState({ rawLinks: links })}
-      getSuggestions={getPanelLinksVariableSuggestions}
-      data={[]}
-    />
-  );
-}
-
-function PanelFrameTitle({ panel }: { panel: VizPanel }) {
-  const { title } = panel.useState();
-
-  return (
-    <Input id="PanelFrameTitle" value={title} onChange={(e) => panel.setState({ title: e.currentTarget.value })} />
-  );
-}
-
-function DescriptionTextArea({ panel }: { panel: VizPanel }) {
-  const { description } = panel.useState();
-
-  return (
-    <TextArea
-      id="description-text-area"
-      value={description}
-      onChange={(e) => panel.setState({ description: e.currentTarget.value })}
-    />
-  );
 }
